+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 import Navbar from './components/Navbar';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import Profile from './pages/Profile/Profile';
+import Profile from './pages/Profile';
 import myCampaigns from './pages/myCampaigns';
 import myEntries from './pages/myEntries';
